refactor: following the checklist
