+++ src/main.tsx
@@ -1,6 +1,4 @@
 import ReactDOM from 'react-dom/client'
-import App from './App.tsx'
+import App from './components/App.tsx'
 
-ReactDOM.createRoot(document.getElementById('root')!).render(
-    <App />
-)
+ReactDOM.createRoot(document.getElementById('root')!).render(<App />)
